Added in Leaflab Maple support
Added in support for the Leaflabs Maple platform and allowed either SPI
port to be used as well as ensuring bit banging still work. This port
also removed AVR optimisations; a branch will be conducted later to
enable both AVR and maple support in a single library as well as AVR &
STM32F103 optimisations.